requirements.txt
appnope==0.1.2
astroid==2.4.2
autopep8==1.5.5
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.17.1
botocore==1.20.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
Faker==5.8.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.1.0
gunicorn==20.0.4
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
SQLAlchemy==1.3.23
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
